+++ src/Plane/Polygons/AbstractPolygon.php
@@ -1,8 +1,17 @@
 <?php
 namespace samizdam\Geometry\Plane\Polygons;
 
+use samizdam\Geometry\Plane\CalculatorAwareTrait;
+
+/**
+ *
+ * @author samizdam
+ *        
+ */
 abstract class AbstractPolygon implements PolygonInterface
 {
+    
+    use CalculatorAwareTrait;
 
     protected $points;