+++ src/Plane/Polygons/PolygonCalculatorInterface.php
@@ -1,0 +1,18 @@
+<?php
+namespace samizdam\Geometry\Plane\Polygons;
+
+/**
+ *
+ * @author samizdam
+ *        
+ */
+interface PolygonCalculatorInterface
+{
+
+    /**
+     * Get area of Polygon
+     *
+     * @param PolygonInterface $polygon
+     */
+    public function getArea(PolygonInterface $polygon);
+}