+++ src/Plane/Polygons/PolygonInspector.php
@@ -1,0 +1,16 @@
+<?php
+namespace samizdam\Geometry\Plane\Polygons;
+
+/**
+ *
+ * @author samizdam
+ *        
+ */
+class PolygonInspector implements PolygonInspectorInterface
+{
+
+    public function isTriangle(PolygonInterface $triangle)
+    {
+        return count($triangle->getPoints()) === 3;
+    }
+}