+++ src/Plane/Polygons/PolygonInspectorInterface.php
@@ -1,0 +1,14 @@
+<?php
+namespace samizdam\Geometry\Plane\Polygons;
+
+
+/**
+ *
+ * @author samizdam
+ *        
+ */
+interface PolygonInspectorInterface
+{
+
+    public function isTriangle(PolygonInterface $triangle);
+}